Move deprecates BladeControl properties to the BladeControl class
